# Vectorize `is_sorted_until`