--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -8092,7 +8092,7 @@ namespace ranges {
         if constexpr (_Is_min_max_iterators_safe<_It> && sized_sentinel_for<_Se, _It> && is_same_v<_Pj, identity>) {
             constexpr bool _Is_greater = _Is_predicate_greater<_It, _Pr>;
             if constexpr (_Is_greater || _Is_predicate_less<_It, _Pr>) {
-                if (!_STD _Is_constant_evaluated()) {
+                if (!_STD is_constant_evaluated()) {
                     const auto _First_ptr = _STD _To_address(_First);
                     const auto _Last_ptr  = _First_ptr + static_cast<size_t>(_Last - _First);
                     const auto _Result    = _STD _Is_sorted_until_vectorized(_First_ptr, _Last_ptr, _Is_greater);