… benchmarks/CMakeLists.txt
… benchmarks/src/is_sorted_until.cpp
… tests/std/include/test_is_sorted_until_support.hpp
… tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
… tests/std/tests/VSO_0000000_vector_algorithms_floats/test.cpp