@AlexGuteniev 2025-04-23
> 5950X results

Interesting thing about floating types in your results. In mine, they were consistently slower, but I suspected "code alignment gremlins", rather than slower instructions itself. The instructions have a bit longer encoding, making stepping into that a bit easier. Your result seem to confirm that, having one of floating results consistent with integer results.