@AlexGuteniev 2025-04-23
> > Your result seem to confirm that, having one of floating results consistent with integer results
> 
> And this happens not because of different CPU, but because a different overload of `ranges::is_sorted` was used.

No, it has changed in tests, not benchmark. Then probably merging re-shuffled things.